03-aidd/pyproject.toml
[project]
name = "aidd-bot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "aiogram>=3.6.0",
  "python-dotenv>=1.0.1",
  "aiohttp-socks>=0.8.4",
]

[tool.ruff]
line-length = 100


